+++ CHANGELOG.md
@@ -31,6 +31,10 @@ speaking, considered to be the player's experience.
   Harbour ship list.
 - Added a "too close" message to Inns when they're too close to other Inns and
   so will have fewer guests.
+- Show the number of units a leader is leading in the leadership status on the
+  info pane.
+- Added clicking on the leadership status on the info pane to go to the location
+  of the leader of the current unit.
 
 ### Fixed
 
@@ -44,6 +48,8 @@ speaking, considered to be the player's experience.
   without a selection change.
 - Draw the building information (builder icon, hit points bar, progress bar, and
   information panel) above the link lines and buttons.
+- Fixed Phoenix not revealing the leadership status (icon and info pane field)
+  of units.
 - Fixed the corrupted display of a Mayan Town name, a Norman unit name, and a
   Greek unit name in the Esperanto, Polish, and Russian translations.
 - Fixed the alignment of the town migration button (pink cross with arrows in a